Combo-specifici hack: Consider a TBR with less than 60 seconds left as completed
when determining whether to set a new TBR.

Setting a new TBR overrides an running TBR if any, so this won't fail on
the Combo. Without this, a TBR mostly runs out 20 seconds after a loop
iteration (when receiving a SGV), so for that almost 5 minutes to TBR
is set.

Note that this is Combo-specific and needs to be revised when
integrating this with mainline.