Show profile switch button in profiles
Shows the button to trigger a profile switch button in the profile fragments even when the pump basal cannot be written.
Instead of "SEND TO PUMP" it will then show "ACTIVATE PROFILE".

This has become necessary with the interval logic that needs profile switch events to happen in order to accept a new profile.