NSTreatments NPE fix
When the dialog was open and the phone goes to sleep/lock screen, Android may garbage collect and options is no longer available. -> NPE

I think as quick fix it is ok to not show a dialog at all. If the user stopped using the dialog and lets the phone sleep for a longer while with the phone going to sleep... he can open it again, if he wants it.