Set Source.PUMP when setting/cancelling TBR.
While technically, the history isn't read, we can be reasonably
sure a TBR was set/cancelled by reading the main menu (and for
TBR starts that's the only way to do so). Reading the history
might cause (more) problems with the pump's minute-precision
and not yield much safety. Besides, command mode is expected
to take over yesterday.

(cherry picked from commit 04fb1c8)