Set a higher hard limit for IC ratio
As written in a comment and in the AAPS thread, I think the new upper hard limit of the IC ratio is to low, for childs but also for me (that's why I found the problem).
I run a profile with IC = 60 g when I do long lasting sports activities, the limit of 40 is much to low - and I'm shure that there are childs which have a higher IC ratio in real life. I propose the value of 100 which is relativly compared not as high as the hard limit MAXISF = 720 mg/dl.