Fix re-enabling the loop plugin from overview. Fixes #861.
Note that enabling a loop plugin is hardcoded against the only
impl currently available LoopPlugin. Support for multiple loop
plugins is incomplete (and not needed at this point), since
there's no LoopInterface, but immediately LoopPlugin (as compared
to APSInterface with OpenAPS* impls).
When adding support, the last activated loop plugin must be remembered
so the correct one is re-enabled.