Don't loop with old BG value, don't loop twice for the same value.
Loop is potentially triggered twice when BG reading sent to NS comes back.
This should also deal with backfilled data coming in, since any previous
reading will be older than 9m, for which DatabaseHelper.actualBg()
returns null.

The previous approach to solve multiple invocations for the same value
added a isNew flag to EventNewBG, but since
DatabaseHelper.scheduleBgChange() drops excessive updates the event
carrying that flag is not guarantued to be delivered, resulting in
missed loop invocations. The approach taken now lets the receiver
of the event fully decide how to deal with it.

Should fix, or at least improve, #901, #671.