BgReading.copyFrom: don't update local data.
Like the 'valid' flag, which is not copied and is not synced with
NS, don't sync sourcePlugin or isFiltered.
Otherwise there might be a race condition if the NS-roundtrip of a
locally received BG updates the sourcePlugin field to null,
thus leaving no values for the loop to loop with.