Update PumpType.java
Ammended the basebasalstep values fro the insight and the combo.

this was preventing AAPS from implementing zero basals where the basalstep was higher than the current running basal rate e.g. basal of 0.06u\hr running and a basalstep 0f 0.1u\hr resulted in no zero temp being enacted because of logic in the loop plugin.