1) Added SMS commands: - SMS DISABLE/STOP to disable SMS Service; - TARGET MEAL/ACTIVITY/HYPO to switch to the standard temp targets; - BOLUS 0.60 MEAL, to do a bolus and set the stadard meal TT (ie 90 mg/dL for 45').
2) Modified the SMS Timeout in a range from 3' to 60'.
This preference is useful for parents that will manage many boluses in a single meal (because often, eaten foods are unpredictable with babies, so parents need to perform many closed boluses).
Due to safety reasons (prevent multiple boluses for stolen phones, cloned SIM, ...):
- Added a further SMS command to disable SMS Services (SMS DISABLE/STOP);
- To modify the default timeout of 15', 2 Phone Numbers are mandatory
In this way, if unwanted/suspicious boluses are notified, the other parent can disable Remote Management (till to re-enable it directly with the AAPS in the master smartphone).