ComboPlugin: fix handling benin warnings.
This change provides usable input to the check* methods. Since a dummy PumpState is returned when a command runs into a warning/alarm on the pump, the checkAndResolveTbrMismatch() reads that no TBR is active and creates a TBR-end record in the DB, which is wrong (and yes, returning a dummy PumpState rather than returning and handling null is a questionable design decision). Effectively this is only a concern when AAPS confirms a low cartridge or low battery alarm, so it's impact is small.