Fix reservoir changes on MDT pumps being logged as cannula changes
When the cannula tubing is primed on Medtronic pumps (manual prime), the history entry that gets logged is the same type as a fixed-prime (for priming the cannula after it has been inserted) so any time the reservoir was replaced AAPS would log a site change which reset autosens unnecessarily and also led to inaccurate reporting of insulin/cannula changes in Nightscout.

This commit aims to fix that by checking the `FixedAmount` data entry of `Prime` history entries to detect whether the prime was a manual-prime (priming tubing) or fixed-prime (priming cannula) and skip logging a site change event if it is the former. In order to automatically log insulin reservoir replacements, this commit also adds a monitor for `Rewind` history entries which logs an `InsulinChange` event for them (similar to the priming monitor).