Prevent NPE when checking for battery level
This is a simple caller-side fix to prevent NullPointerException when checking battery level. The NPE happened  when the BT stack returned an empty array in the response with a `RESULT_SUCCESS` result code.

We now check for empty array, and log that as a seperate error is such a response is received.