Allow/reenable the setting of TempTargets while pump is suspended or disconnected
Unfortunately, it is currently no longer possible to set a TempTarget when the pump is disconnected or suspended (TBR: 0%). This seems logical at first sight (why a TempTarget when the pump is not connected or the loop suspended!), but prevents many useful possibilities, namely when the TempTarget should be set immediately after the disconnection or suspension has expired.
An example of this are f. e. sport activities: To reduce IOB, I set a 0% TBR half an hour before sport for one hour with the function "pump disconnected", so that the associated 0% TBR continues for a while while sport starts. I also set a three hour Activity TempTarget. This continues to be active when the pump loops again with the pump disconnected time expired.
The fix allows/reenables setting a TempTarget even if the pump is suspended or disconnected.