fix: don't restart alarm sound service automatically
This sets the alarm sound service to not be sticky. If AAPS is stopped while an alarm is playing, the alarm sound service will be started again by the system but there's no built-in way to stop it after this happens. This commit makes it so the alarm sound service won't be restarted by Android, thereby avoiding the aforementioned issue.