combov2: Check if pump is suspended before setting TBR and delivering bolus
This avoids a misleading error state caused by an IllegalStateException
that gets thrown by the Pump.setTbr() and Pump.deliverBolus() when these
get called while the pump is suspended. Instead, first check for this
suspended state, and if the pump is indeed suspended, immediately exit
with the pump enact result set to not have succeeded nor have been enacted.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>