combov2: Use dedicated variable for keeping track of suspended state
This is necessary to prevent race conditions during connection setup.
isSuspended() was relying on the driverStateUIFlow to determine if the
pump is currently suspended. However, because driverStateUIFlow is updated
in a separate coroutine, there is a chance that the Suspended state is
set shortly _after_ connect() called isSuspended() to determine if the
pump is suspended or not, causing connect() to do the wrong thing and
assume that the pump is currently connected. Since the main screen does
not show the current profile number, this leads to the error message
"profile 1 must be the active profile, not profile 0".